Refactor request IP and ID retrievals in server handling
This commit updates the server.js to improve the way it retrieves request IPs and IDs. Instead of fetching this information in several lines of code across the functions, the refactor adopts a more efficient, streamlined approach by assigning these values to variables and reusing them. This change reduces code redundancy and enhances readability.

Signed-off-by: Akama Aka <akama.aka@akami-solutions.cc>